Add VBS_DESCRIPTOR support to XIC-RPC framework. Add DemoClientImp.cpp and DemoClientImp.h Update the document vbs.rst and xic.rst Change the item syntax of section verifier in shadow file.

+++ xic/VWriter.h
@@ -65,7 +65,7 @@ protected:
 		void close();
 		void finish_child();
 
-		void writeDescriptor(int16_t descriptor) { vbs_pack_descriptor(_pk, descriptor); }
+		void writeDescriptor(int descriptor) { vbs_pack_descriptor(_pk, descriptor); }
 
 		void raw(const void *buf, size_t size)	{ vbs_pack_raw(_pk, buf, size); }
 
@@ -147,37 +147,37 @@ public:
 	void flush();
 
 	template <typename ValueType>
-	void v(const ValueType& v, int16_t descriptor=0);
+	void v(const ValueType& v, int descriptor=0);
 
 	void vfmt(xfmt_callback_function cb, const char *fmt, ...);
 	void vfmt(xfmt_callback_function cb, const char *fmt, va_list ap);
 
-	void vstring(const xstr_t& xstr, int16_t descriptor=0);
-	void vstring(const char *str, size_t len, int16_t descriptor=0);
-	void vstring(const rope_t *rope, int16_t descriptor=0);
-	void vstring(const struct iovec *iov, int count, int16_t descriptor=0);
+	void vstring(const xstr_t& xstr, int descriptor=0);
+	void vstring(const char *str, size_t len, int descriptor=0);
+	void vstring(const rope_t *rope, int descriptor=0);
+	void vstring(const struct iovec *iov, int count, int descriptor=0);
 
-	void vblob(const xstr_t& blob, int16_t descriptor=0);
-	void vblob(const void *data, size_t len, int16_t descriptor=0);
-	void vblob(const rope_t *rope, int16_t descriptor=0);
-	void vblob(const struct iovec *iov, int count, int16_t descriptor=0);
+	void vblob(const xstr_t& blob, int descriptor=0);
+	void vblob(const void *data, size_t len, int descriptor=0);
+	void vblob(const rope_t *rope, int descriptor=0);
+	void vblob(const struct iovec *iov, int count, int descriptor=0);
 
-	void vstanza(const unsigned char *buf, size_t size, int16_t descriptor=0);
+	void vstanza(const unsigned char *buf, size_t size, int descriptor=0);
 
 	void vnull();
 
-	VListWriter vlist(int16_t descriptor=0);
+	VListWriter vlist(int descriptor=0);
 
 	template <typename ListType>
-	VListWriter vlist(const ListType& values, int16_t descriptor=0);
+	VListWriter vlist(const ListType& values, int descriptor=0);
 
-	VDictWriter vdict(int16_t descriptor=0);
+	VDictWriter vdict(int descriptor=0);
 
 	template <typename DictType>
-	VDictWriter vdict(const DictType& dict, int16_t descriptor=0);
+	VDictWriter vdict(const DictType& dict, int descriptor=0);
 
-	void vstrhead(size_t len, int16_t descriptor=0);
-	void vblobhead(size_t len, int16_t descriptor=0);
+	void vstrhead(size_t len, int descriptor=0);
+	void vblobhead(size_t len, int descriptor=0);
 
 	// Called after vstrhead() or vblobhead();
 	void raw(const void *buf, size_t size);
@@ -196,7 +196,7 @@ public:
 	void flush();
 
 	template <typename KeyType, typename ValueType>
-	void kv(const KeyType& key, const ValueType& value, int16_t descriptor=0);
+	void kv(const KeyType& key, const ValueType& value, int descriptor=0);
 
 	template <typename KeyType>
 	void kvfmt(const KeyType& key, xfmt_callback_function cb, const char *fmt, ...);
@@ -205,52 +205,52 @@ public:
 	void kvfmt(const KeyType& key, xfmt_callback_function cb, const char *fmt, va_list ap);
 
 	template <typename KeyType>
-	void kvstring(const KeyType& key, const xstr_t& xstr, int16_t descriptor=0);
+	void kvstring(const KeyType& key, const xstr_t& xstr, int descriptor=0);
 
 	template <typename KeyType>
-	void kvstring(const KeyType& key, const char *str, size_t len, int16_t descriptor=0);
+	void kvstring(const KeyType& key, const char *str, size_t len, int descriptor=0);
 
 	template <typename KeyType>
-	void kvstring(const KeyType& key, const rope_t *rope, int16_t descriptor=0);
+	void kvstring(const KeyType& key, const rope_t *rope, int descriptor=0);
 
 	template <typename KeyType>
-	void kvstring(const KeyType& key, const struct iovec *iov, int count, int16_t descriptor=0);
+	void kvstring(const KeyType& key, const struct iovec *iov, int count, int descriptor=0);
 
 	template <typename KeyType>
-	void kvblob(const KeyType& key, const xstr_t& blob, int16_t descriptor=0);
+	void kvblob(const KeyType& key, const xstr_t& blob, int descriptor=0);
 
 	template <typename KeyType>
-	void kvblob(const KeyType& key, const void *data, size_t len, int16_t descriptor=0);
+	void kvblob(const KeyType& key, const void *data, size_t len, int descriptor=0);
 
 	template <typename KeyType>
-	void kvblob(const KeyType& key, const rope_t *rope, int16_t descriptor=0);
+	void kvblob(const KeyType& key, const rope_t *rope, int descriptor=0);
 
 	template <typename KeyType>
-	void kvblob(const KeyType& key, const struct iovec *iov, int count, int16_t descriptor=0);
+	void kvblob(const KeyType& key, const struct iovec *iov, int count, int descriptor=0);
 
 	template <typename KeyType>
-	void kvstanza(const KeyType& key, const unsigned char *buf, size_t size, int16_t descriptor=0);
+	void kvstanza(const KeyType& key, const unsigned char *buf, size_t size, int descriptor=0);
 
 	template <typename KeyType>
 	void kvnull(const KeyType& key);
 
 	template <typename KeyType>
-	VListWriter kvlist(const KeyType& key, int16_t descriptor=0);
+	VListWriter kvlist(const KeyType& key, int descriptor=0);
 
 	template <typename KeyType, typename ListType>
-	VListWriter kvlist(const KeyType& key, const ListType& values, int16_t descriptor=0);
+	VListWriter kvlist(const KeyType& key, const ListType& values, int descriptor=0);
 
 	template <typename KeyType>
-	VDictWriter kvdict(const KeyType& key, int16_t descriptor=0);
+	VDictWriter kvdict(const KeyType& key, int descriptor=0);
 
 	template <typename KeyType, typename DictType>
-	VDictWriter kvdict(const KeyType& key, const DictType& dict, int16_t descriptor=0);
+	VDictWriter kvdict(const KeyType& key, const DictType& dict, int descriptor=0);
 
 	template <typename KeyType>
-	void kvstrhead(const KeyType& key, size_t len, int16_t descriptor=0);
+	void kvstrhead(const KeyType& key, size_t len, int descriptor=0);
 
 	template <typename KeyType>
-	void kvblobhead(const KeyType& key, size_t len, int16_t descriptor=0);
+	void kvblobhead(const KeyType& key, size_t len, int descriptor=0);
 
 	// Called after kvstrhead() or kvblobhead();
 	void raw(const void *buf, size_t size);
@@ -265,7 +265,7 @@ inline void VListWriter::flush()
 }
 
 template <typename ValueType>
-inline void VListWriter::v(const ValueType& v, int16_t descriptor)
+inline void VListWriter::v(const ValueType& v, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -292,7 +292,7 @@ inline void VListWriter::vfmt(xfmt_callback_function cb, const char *fmt, va_lis
 	_entity->writeFormat(cb, fmt, ap);
 }
 
-inline void VListWriter::vstring(const xstr_t& xstr, int16_t descriptor)
+inline void VListWriter::vstring(const xstr_t& xstr, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -301,7 +301,7 @@ inline void VListWriter::vstring(const xstr_t& xstr, int16_t descriptor)
 	_entity->write(xstr);
 }
 
-inline void VListWriter::vstring(const char *str, size_t len, int16_t descriptor)
+inline void VListWriter::vstring(const char *str, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -310,7 +310,7 @@ inline void VListWriter::vstring(const char *str, size_t len, int16_t descriptor
 	_entity->writeString(str, len);
 }
 
-inline void VListWriter::vstring(const rope_t *rope, int16_t descriptor)
+inline void VListWriter::vstring(const rope_t *rope, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -319,7 +319,7 @@ inline void VListWriter::vstring(const rope_t *rope, int16_t descriptor)
 	_entity->writeString(rope);
 }
 
-inline void VListWriter::vstring(const struct iovec *iov, int count, int16_t descriptor)
+inline void VListWriter::vstring(const struct iovec *iov, int count, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -328,7 +328,7 @@ inline void VListWriter::vstring(const struct iovec *iov, int count, int16_t des
 	_entity->writeString(iov, count);
 }
 
-inline void VListWriter::vblob(const xstr_t& blob, int16_t descriptor)
+inline void VListWriter::vblob(const xstr_t& blob, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -337,7 +337,7 @@ inline void VListWriter::vblob(const xstr_t& blob, int16_t descriptor)
 	_entity->writeBlob(blob.data, blob.len);
 }
 
-inline void VListWriter::vblob(const void *data, size_t len, int16_t descriptor)
+inline void VListWriter::vblob(const void *data, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -346,7 +346,7 @@ inline void VListWriter::vblob(const void *data, size_t len, int16_t descriptor)
 	_entity->writeBlob(data, len);
 }
 
-inline void VListWriter::vblob(const rope_t *rope, int16_t descriptor)
+inline void VListWriter::vblob(const rope_t *rope, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -355,7 +355,7 @@ inline void VListWriter::vblob(const rope_t *rope, int16_t descriptor)
 	_entity->writeBlob(rope);
 }
 
-inline void VListWriter::vblob(const struct iovec *iov, int count, int16_t descriptor)
+inline void VListWriter::vblob(const struct iovec *iov, int count, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -364,7 +364,7 @@ inline void VListWriter::vblob(const struct iovec *iov, int count, int16_t descr
 	_entity->writeBlob(iov, count);
 }
 
-inline void VListWriter::vstanza(const unsigned char *buf, size_t size, int16_t descriptor)
+inline void VListWriter::vstanza(const unsigned char *buf, size_t size, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -380,7 +380,7 @@ inline void VListWriter::vnull()
 	return _entity->writeNull();
 }
 
-inline VListWriter VListWriter::vlist(int16_t descriptor)
+inline VListWriter VListWriter::vlist(int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -390,7 +390,7 @@ inline VListWriter VListWriter::vlist(int16_t descriptor)
 }
 
 template <typename ListType>
-inline VListWriter VListWriter::vlist(const ListType& values, int16_t descriptor)
+inline VListWriter VListWriter::vlist(const ListType& values, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -402,7 +402,7 @@ inline VListWriter VListWriter::vlist(const ListType& values, int16_t descriptor
 	return lw;
 }
 
-inline VDictWriter VListWriter::vdict(int16_t descriptor)
+inline VDictWriter VListWriter::vdict(int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -412,7 +412,7 @@ inline VDictWriter VListWriter::vdict(int16_t descriptor)
 }
 
 template <typename DictType>
-inline VDictWriter VListWriter::vdict(const DictType& dict, int16_t descriptor)
+inline VDictWriter VListWriter::vdict(const DictType& dict, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -425,7 +425,7 @@ inline VDictWriter VListWriter::vdict(const DictType& dict, int16_t descriptor)
 }
 
 
-inline void VListWriter::vstrhead(size_t len, int16_t descriptor)
+inline void VListWriter::vstrhead(size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -434,7 +434,7 @@ inline void VListWriter::vstrhead(size_t len, int16_t descriptor)
 	_entity->writeStrHead(len);
 }
 
-inline void VListWriter::vblobhead(size_t len, int16_t descriptor)
+inline void VListWriter::vblobhead(size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -459,7 +459,7 @@ inline void VDictWriter::flush()
 }
 
 template <typename KeyType, typename ValueType>
-inline void VDictWriter::kv(const KeyType& key, const ValueType& value, int16_t descriptor)
+inline void VDictWriter::kv(const KeyType& key, const ValueType& value, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -488,7 +488,7 @@ inline void VDictWriter::kvfmt(const KeyType& key, xfmt_callback_function cb, co
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstring(const KeyType& key, const xstr_t& xstr, int16_t descriptor)
+inline void VDictWriter::kvstring(const KeyType& key, const xstr_t& xstr, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -499,7 +499,7 @@ inline void VDictWriter::kvstring(const KeyType& key, const xstr_t& xstr, int16_
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstring(const KeyType& key, const char *str, size_t len, int16_t descriptor)
+inline void VDictWriter::kvstring(const KeyType& key, const char *str, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -510,7 +510,7 @@ inline void VDictWriter::kvstring(const KeyType& key, const char *str, size_t le
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstring(const KeyType& key, const rope_t *rope, int16_t descriptor)
+inline void VDictWriter::kvstring(const KeyType& key, const rope_t *rope, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -521,7 +521,7 @@ inline void VDictWriter::kvstring(const KeyType& key, const rope_t *rope, int16_
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstring(const KeyType& key, const struct iovec *iov, int count, int16_t descriptor)
+inline void VDictWriter::kvstring(const KeyType& key, const struct iovec *iov, int count, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -532,7 +532,7 @@ inline void VDictWriter::kvstring(const KeyType& key, const struct iovec *iov, i
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvblob(const KeyType& key, const xstr_t& blob, int16_t descriptor)
+inline void VDictWriter::kvblob(const KeyType& key, const xstr_t& blob, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -543,7 +543,7 @@ inline void VDictWriter::kvblob(const KeyType& key, const xstr_t& blob, int16_t 
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvblob(const KeyType& key, const void *data, size_t len, int16_t descriptor)
+inline void VDictWriter::kvblob(const KeyType& key, const void *data, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -554,7 +554,7 @@ inline void VDictWriter::kvblob(const KeyType& key, const void *data, size_t len
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvblob(const KeyType& key, const rope_t *rope, int16_t descriptor)
+inline void VDictWriter::kvblob(const KeyType& key, const rope_t *rope, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -565,7 +565,7 @@ inline void VDictWriter::kvblob(const KeyType& key, const rope_t *rope, int16_t 
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvblob(const KeyType& key, const struct iovec *iov, int count, int16_t descriptor)
+inline void VDictWriter::kvblob(const KeyType& key, const struct iovec *iov, int count, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -576,7 +576,7 @@ inline void VDictWriter::kvblob(const KeyType& key, const struct iovec *iov, int
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstanza(const KeyType& key, const unsigned char *buf, size_t size, int16_t descriptor)
+inline void VDictWriter::kvstanza(const KeyType& key, const unsigned char *buf, size_t size, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -596,7 +596,7 @@ inline void VDictWriter::kvnull(const KeyType& key)
 }
 
 template <typename KeyType>
-inline VListWriter VDictWriter::kvlist(const KeyType& key, int16_t descriptor)
+inline VListWriter VDictWriter::kvlist(const KeyType& key, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -607,7 +607,7 @@ inline VListWriter VDictWriter::kvlist(const KeyType& key, int16_t descriptor)
 }
 
 template <typename KeyType, typename ListType>
-inline VListWriter VDictWriter::kvlist(const KeyType& key, const ListType& values, int16_t descriptor)
+inline VListWriter VDictWriter::kvlist(const KeyType& key, const ListType& values, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -622,7 +622,7 @@ inline VListWriter VDictWriter::kvlist(const KeyType& key, const ListType& value
 
 
 template <typename KeyType>
-inline VDictWriter VDictWriter::kvdict(const KeyType& key, int16_t descriptor)
+inline VDictWriter VDictWriter::kvdict(const KeyType& key, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -633,7 +633,7 @@ inline VDictWriter VDictWriter::kvdict(const KeyType& key, int16_t descriptor)
 }
 
 template <typename KeyType, typename DictType>
-inline VDictWriter VDictWriter::kvdict(const KeyType& key, const DictType& dict, int16_t descriptor)
+inline VDictWriter VDictWriter::kvdict(const KeyType& key, const DictType& dict, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -647,7 +647,7 @@ inline VDictWriter VDictWriter::kvdict(const KeyType& key, const DictType& dict,
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvstrhead(const KeyType& key, size_t len, int16_t descriptor)
+inline void VDictWriter::kvstrhead(const KeyType& key, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
@@ -658,7 +658,7 @@ inline void VDictWriter::kvstrhead(const KeyType& key, size_t len, int16_t descr
 }
 
 template <typename KeyType>
-inline void VDictWriter::kvblobhead(const KeyType& key, size_t len, int16_t descriptor)
+inline void VDictWriter::kvblobhead(const KeyType& key, size_t len, int descriptor)
 {
 	if (_entity->child)
 		_entity->finish_child();
